+++ backend-opal/src/server/__mocks__/stockServerTestData.js
@@ -1,0 +1,8 @@
+const stockItems = [
+  { id: 1, userID: 1, symbol: 'A', amount: 10 },
+  { id: 2, userID: 2, symbol: 'B', amount: 20 },
+  { id: 3, userID: 3, symbol: 'C', amount: 30 },
+  { id: 4, userID: 4, symbol: 'D', amount: 40 }
+];
+
+module.exports = { stockItems };
