--- backend-opal/src/server/__mocks__/stockTestData.js
@@ -1,8 +1,0 @@
-const stockItems = [
-  { id: 1, userID: 1, symbol: 'MSFT', amount: 100 },
-  { id: 2, userID: 2, symbol: 'AMZN', amount: 200 },
-  { id: 3, userID: 3, symbol: 'AAPL', amount: 300 },
-  { id: 4, userID: 4, symbol: 'FB', amount: 400 }
-];
-
-module.exports = { stockItems };
